+++ code/proof-methodologies/lemma-functions/insert-sort-auto.c
@@ -1,5 +1,5 @@
 #include <stddef.h>
-#include <limits.h>
+#include <stdint.h>
 
 /*@
   axiomatic Occurrences_Axiomatic{
@@ -48,7 +48,7 @@
 */
 
 /*@
-  requires beg < last < UINT_MAX && \valid(a + (beg .. last));
+  requires beg < last < SIZE_MAX && \valid(a + (beg .. last));
   requires sorted(a, beg, last) ;
 
   assigns a[ beg .. last ] ;
@@ -99,7 +99,7 @@ void insertion_sort(int* a, size_t beg, size_t end){
     loop variant end-i ;
   */
   for(size_t i = beg+1; i < end; ++i) {
-    //@ ghost L:
+    //@ ghost L:;
     insert(a, beg, i);
     //@ assert permutation{L, Here}(a, beg, i+1);
     //@ assert unchanged{L, Here}(a, i+1, end) ;
